Added docker-compose with local kafka cluster Aligned expected records with records generated by Flink DataGenerator example
