+++ java/KafkaConnectors/src/main/resources/flink-application-properties-dev.json
@@ -2,28 +2,16 @@
   {
     "PropertyGroupId": "Input0",
     "PropertyMap": {
-      "bootstrap.servers": "<BootstrapServers>",
-      "topic": "<Topic>",
-      "group.id": "<ConsumerGroupID>"
+      "bootstrap.servers": "localhost:9092",
+      "topic": "stock-prices",
+      "group.id": "flink-kafka-example"
     }
   },
   {
     "PropertyGroupId": "Output0",
     "PropertyMap": {
-      "bootstrap.servers": "<BootstrapServers>",
-      "topic": "<Topic>",
-      "transaction.timeout.ms": "1000"
-    }
-  },
-  {
-    "PropertyGroupId": "AuthProperties",
-    "PropertyMap": {
-      "sasl.mechanism": "AWS_MSK_IAM",
-      "sasl.client.callback.handler.class": "software.amazon.msk.auth.iam.IAMClientCallbackHandler",
-      "sasl.jaas.config": "software.amazon.msk.auth.iam.IAMLoginModule required;",
-      "security.protocol": "SASL_SSL",
-      "ssl.truststore.location": "/usr/lib/jvm/java-11-amazon-corretto/lib/security/cacerts",
-      "ssl.truststore.password": "changeit"
+      "bootstrap.servers": "localhost:9092",
+      "topic": "stock-prices-out"
     }
   }
 ]
